Devolviendo 204 No Content y 201 Created en POST y DELETE

+++ proyectoMaximoSpringBoot/src/main/java/com/proyectoMaximo/proyectoMaximoSpringBoot/servicios/ServicioService.java
@@ -1,10 +1,13 @@
 package com.proyectoMaximo.proyectoMaximoSpringBoot.servicios;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import java.util.List;
 
@@ -13,13 +16,35 @@ public class ServicioService {
 
     @Autowired ServicioRepository servicioRepository;
 
-    public void registrarServicio(DatosRegistroServicios datos) {
+    public ResponseEntity<DatosRespuestaServicio> registrarServicio(
+            DatosRegistroServicios datos, UriComponentsBuilder uriComponentsBuilder) {
 
         if (servicioRepository.existsByCodigoServicio(datos.codigoServicio())) {
             throw new RuntimeException("Código de servicio duplicado");
         }
+        Servicio servicio = new Servicio(datos);//Creamos el objeto
         servicioRepository.save(new Servicio(datos));
+
+        //Construímos la uri del recurso creado
+        var uri = uriComponentsBuilder.path(
+                "/servicios/{id}").buildAndExpand(servicio.getId()).toUri();
+
+        //Creamos la respuesta
+        DatosRespuestaServicio datosRespuestaServicio = new DatosRespuestaServicio(
+                servicio.getId(),
+                servicio.getCodigoServicio(),
+                servicio.getItem(),
+                servicio.getAgrupacion(),
+                servicio.getEstructura(),
+                servicio.getDetalle(),
+                servicio.getUnidad(),
+                servicio.getValorUnitario(),
+                servicio.getNivelTension(),
+                servicio.getDescripcion()
+        );
+        return ResponseEntity.created(uri).body(datosRespuestaServicio);
     }
+
     public Page<DatosListadoServicios> listarServicios(Pageable paginacion) {
         return servicioRepository
                 .findAll(paginacion)
@@ -27,8 +52,9 @@ public class ServicioService {
     }
 
     @Transactional
-    public void actualizarServicio(DatosActualizarServicios datos) {
+    public ResponseEntity actualizarServicio(DatosActualizarServicios datos) {
         Servicio servicio = servicioRepository.getReferenceById(datos.id());
+        servicio.actualizarDatos(datos);
 
         if (datos.codigoServicio() != null) servicio.setCodigoServicio(datos.codigoServicio());
         if (datos.item() != null) servicio.setItem(datos.item());
@@ -39,9 +65,34 @@ public class ServicioService {
         if (datos.valorUnitario() != null) servicio.setValorUnitario(datos.valorUnitario());
         if (datos.nivelTension() != null) servicio.setNivelTension(datos.nivelTension());
         if (datos.descripcion() != null) servicio.setDescripcion(datos.descripcion());
+
+        return ResponseEntity.ok(new DatosRespuestaServicio(
+                servicio.getId(),
+                servicio.getCodigoServicio(),
+                servicio.getItem(),
+                servicio.getAgrupacion(),
+                servicio.getEstructura(),
+                servicio.getDetalle(),
+                servicio.getUnidad(),
+                servicio.getValorUnitario(),
+                servicio.getNivelTension(),
+                servicio.getDescripcion()
+        ));
     }
 
     public void eliminarServicio(Long id) {
+
+        if (!servicioRepository.existsById(id)) {
+            throw new EntityNotFoundException("Servicio no encontrado con el id: " + id);
+        }
         servicioRepository.deleteById(id);
+
     }
+
+    /*public ResponseEntity eliminarServicio(Long id) {
+        servicioRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Servicio no encontrado"));
+
+        return ResponseEntity.noContent().build();
+    }*/
 }
